API: Fixed issue with checking OS rules for unsupported OS, arch or OS version
While tinkering on Termux (android), I stumbled upon an annoying issue where Windows or macOS arguments were bypassing an OS rule. This was due to the filter being completely ignored if the name, os or version were not successfully determined. We are now rejecting if such variables could not be determined.

WORTH NOTING that I'm still stuck on Termux, trying to recompile LWJGL from source, but it's now failing at link time due to libffi being compiled for glibc and not android libc...